evalakarata

+++ src/org/webserver/PutHttpCall.java
@@ -4,28 +4,39 @@ import java.io.File;
 import java.io.IOException;
 
 public class PutHttpCall implements HttpCall {
-	
-	Validator validator;
-	
+
+	private static Validator validator;
+
+	static {
+		validator = new Validator(){
+			
+		};
+	}
+
 	@Override
 	public void execute(HttpRequest req, HttpResponse resp) {
+
 
 		String uri = req.getUri();
 
 		String filePath = FILESYSTEM + uri;
 		File file = new File(filePath);
 
+		if(!validator.validateFile(filePath)){
+			resp.sendError(HTTPCode.ERROR);
+			return;
+		}
+		
 		try {
 
 			boolean createdFile = file.createNewFile();
-			System.out.println(createdFile);
-			
+
 			if (!createdFile) {
 				file.delete();
 				file.createNewFile();
 			}
 			resp.sendOK();
-			
+
 		} catch (IOException e) {
 			e.printStackTrace();
 			resp.sendError(HTTPCode.ERROR);
